requirements.txt
python_telegram_bot==12.5.1
selenium==3.9.1
python-dotenv==0.19.0
  